move SamplingRulesFormat
