+++ tracer/src/Datadog.Trace/Sampling/SamplingRulesFormat.cs
@@ -1,13 +1,13 @@
-// <copyright file="CustomSamplingRulesFormat.cs" company="Datadog">
+// <copyright file="SamplingRulesFormat.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
 
 #nullable enable
 
-namespace Datadog.Trace.Configuration;
+namespace Datadog.Trace.Sampling;
 
-internal static class CustomSamplingRulesFormat
+internal static class SamplingRulesFormat
 {
     public const string Regex = "regex";
